# src/main/java/com/example/bankcards/dto/CardBalance.java
package com.example.bankcards.dto;

import lombok.*;

import java.math.BigDecimal;

@Getter
@Setter
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class CardBalance {
    private BigDecimal balance;
}